requirements.txt
numpy==1.24.4
opencv-python-headless
Pillow==10.4.0
fastapi==0.115.0
uvicorn==0.31.0
gunicorn==23.0.0
python-multipart==0.0.12